StreamLit_Snowflake_Report/requirements.txt
streamlit
pandas
numpy
snowflake-connector-python
python-doten